btc_study/requirements.txt
ccxt
python-dotenv
requests